+++ Rdmp.UI/SimpleDialogs/UserSettingsUI.Designer.cs
@@ -78,6 +78,7 @@
             this.userSettingsToolTips = new System.Windows.Forms.ToolTip(this.components);
             this.cbSelectiveRefresh = new System.Windows.Forms.CheckBox();
           
+            this.cbAlwaysJoinEverything = new System.Windows.Forms.CheckBox();
             this.groupBox1.SuspendLayout();
             ((System.ComponentModel.ISupportInitialize)(this.olvErrorCodes)).BeginInit();
             this.groupBox2.SuspendLayout();
@@ -514,6 +515,7 @@
             this.groupBox7.Controls.Add(this.ddWordWrap);
             this.groupBox7.Controls.Add(this.label4);
             this.groupBox7.Controls.Add(this.label7);
+            this.groupBox7.Controls.Add(this.cbAlwaysJoinEverything);
             this.groupBox7.Controls.Add(this.cbShowCohortWizard);
             this.groupBox7.Location = new System.Drawing.Point(3, 345);
             this.groupBox7.Name = "groupBox7";
@@ -605,6 +607,18 @@
             this.cbSelectiveRefresh.Text = "Selective Refresh";
             this.cbSelectiveRefresh.UseVisualStyleBackColor = true;
             this.cbSelectiveRefresh.CheckedChanged += new System.EventHandler(this.cb_CheckedChanged);
+            //
+            // cbAlwaysJoinEverything
+            // 
+            this.cbAlwaysJoinEverything.AutoSize = true;
+            this.cbAlwaysJoinEverything.Location = new System.Drawing.Point(8, 184);
+            this.cbAlwaysJoinEverything.Margin = new System.Windows.Forms.Padding(4, 3, 4, 3);
+            this.cbAlwaysJoinEverything.Name = "cbAlwaysJoinEverything";
+            this.cbAlwaysJoinEverything.Size = new System.Drawing.Size(146, 19);
+            this.cbAlwaysJoinEverything.TabIndex = 2;
+            this.cbAlwaysJoinEverything.Text = "Always Join Everything";
+            this.cbAlwaysJoinEverything.UseVisualStyleBackColor = true;
+            this.cbAlwaysJoinEverything.CheckedChanged += new System.EventHandler(this.cb_CheckedChanged);
             // 
             // UserSettingsFileUI
             // 
@@ -690,5 +704,6 @@
         private System.Windows.Forms.TextBox tbArchiveTriggerTimeout;
         private System.Windows.Forms.Label label13;
         private System.Windows.Forms.CheckBox cbSelectiveRefresh;
+        private System.Windows.Forms.CheckBox cbAlwaysJoinEverything;
     }
 }